Random testing of bsg_fifo_1rw_large
